backend/requirements.txt
#backend/requirements.txt
fastapi==0.128.5
uvicorn[standard]==0.40.0
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
python-dotenv==1.0.1
python-multipart==0.0.9
google-auth==2.43.0
alembic==1.13.1
openai==1.108.1
